requirements.txt
alembic==1.10.3
attrs==22.2.0
bcrypt==3.2.0
certifi==2023.5.7
cffi==1.15.1
click==8.0.4
colorama==0.4.6
cryptography==38.0.4
Flask==2.2.2
Flask-Cors==3.0.10
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
greenlet==2.0.1
importlib-metadata==6.6.0
importlib-resources==5.12.0
itsdangerous==2.0.1
Jinja2==3.1.2
jsonschema==4.17.3
Mako==1.2.4
MarkupSafe==2.1.1
pip==23.0.1
pkgutil_resolve_name==1.3.10
pycparser==2.21
PyMySQL==1.0.2
pyrsistent==0.18.0
pyserial==3.5
python-dotenv==0.21.0
regex==2022.7.9
setuptools==65.6.3
six==1.16.0
SQLAlchemy==1.4.39
typing_extensions==4.5.0
Werkzeug==2.2.2
wheel==0.37.1
wincertstore==0.2
zipp==3.15.0
